pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "todo_django"
version = "0.1.0"
description = "The backend for my Todo app project"
authors = [{ name = "Izzat Z.", email = "izzat.zainir11@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "django==5.0.0",
    "djangorestframework==3.14.0",
    "django-filter==23.5",
]

[project.optional-dependencies]
prod = ["psycopg2==2.9.9"]
dev = ["psycopg2-binary==2.9.9", "pre-commit==3.6.0"]
dev-lint = ["ruff==0.1.13", "mypy==1.8.0"]
dev-typing = ["django-stubs>=4.2"]

[tool.ruff]
src = ["src"]
fix = true
unsafe-fixes = true
show-fixes = true
extend-select = [
    "RUF", # Ruff
    "PL",  # PyLint
    "N",   # pep8-naming
    "E",   # pycodestyle error
    "W",   # pycodestyle warning
    "I",   # isort
    "UP",  # pyupgrade
    "S",   # flake8-bandit
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "FA",  # flake8-future-annotations
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
]
extend-exclude = ["**/migrations/*"]

[tool.ruff.isort]
case-sensitive = true
known-first-party = ["todo_django"]

[tool.mypy]
mypy_path = "src"
strict = true
show_error_codes = true
explicit_package_bases = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = ["*.api.*"]
disable_error_code = ["misc"]

[tool.django-stubs]
django_settings_module = "todo_django.app.settings"
